+++ src/main.js
@@ -5,10 +5,6 @@ import App from './components/App/App'
 
 /* eslint-disable no-unused-vars */
 
-// vue-router used to SEO
-import VueRouter from 'vue-router'
-Vue.use(VueRouter)
-
 // vue-material is used for md-tooltips, which display
 // calculator variable info
 import VueMaterial from 'vue-material'
@@ -50,33 +46,15 @@ Vue.component('calc-var-string', CalcVarString)
 import CalcVarCheckbox from 'src/misc/CalculatorEngineV2/view/CalcVarCheckbox.vue'
 Vue.component('calc-var-checkbox', CalcVarCheckbox)
 
+// =========================================== //
+// ============= IMPORT VUEX STORE =========== //
+// =========================================== //
 import store from './store'
 
-// 1. Define route components.
-// These can be imported from other files
-const Foo = { template: '<div>foo</div>' }
-const Bar = { template: '<div>bar</div>' }
-
-// 2. Define some routes
-// Each route should map to a component. The "component" can
-// either be an actual component constructor created via
-// Vue.extend(), or just a component options object.
-// We'll talk about nested routes later.
-const routes = [
-  { path: '/foo', component: Foo },
-  { path: '/bar', component: Bar }
-]
-
-// 3. Create the router instance and pass the `routes` option
-// You can pass in additional options here, but let's
-// keep it simple for now.
-const router = new VueRouter({
-  // Removes the "#" from the URL, making them look "normal". This also requires server configuration to
-  // work correctly (e.g. a .htaccess file for an Apache server). Server config is located in <repo dir>/server-config
-  mode: 'history',
-  routes // short for routes: routes
-})
-
+// =========================================== //
+// ====== IMPORT ROUTER, SYNC WITH STORE ===== //
+// =========================================== //
+import { router } from './router'
 import { sync } from 'vuex-router-sync'
 sync(store, router)